Phase 1 refactoring done
