+++ forui/lib/src/widgets/tile/tile_content.dart
@@ -1,12 +1,13 @@
+import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
+import 'package:forui/src/foundation/rendering.dart';
 
 import 'package:meta/meta.dart';
 
 import 'package:forui/forui.dart';
 import 'package:forui/src/widgets/tile/tile_group.dart';
-import 'package:forui/src/widgets/tile/tile_render_object.dart';
 
 part 'tile_content.style.dart';
 
@@ -29,8 +30,6 @@ class FTileContent extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final ltr = Directionality.maybeOf(context) == TextDirection.ltr;
-
     final tile = FTileData.maybeOf(context)!;
     final FTileData(style: tileStyle, :states, :pressable) = tile;
 
@@ -44,31 +43,24 @@ class FTileContent extends StatelessWidget {
       true => group.divider,
     };
 
-    return TileRenderObject(
+    return _TileContent(
       style: contentStyle,
       divider: dividerType,
-      first: tile.index == 0 && group.index == 0,
-      last: tile.last && group.index == group.length - 1,
-      // We use the left side of the border to draw the focused outline.
-      side: states.contains(WidgetState.focused) ? stateStyle.border.resolve(states).left : null,
       children: [
         if (prefixIcon case final prefix?)
           Padding(
-            padding: ltr
-                ? EdgeInsets.only(right: contentStyle.prefixIconSpacing)
-                : EdgeInsets.only(left: contentStyle.prefixIconSpacing),
+            padding: EdgeInsetsDirectional.only(end: contentStyle.prefixIconSpacing),
             child: IconTheme(data: contentStyle.prefixIconStyle.resolve(states), child: prefix),
           )
         else
           const SizedBox(),
         Padding(
-          padding: ltr
-              ? EdgeInsets.only(right: contentStyle.middleSpacing)
-              : EdgeInsets.only(left: contentStyle.middleSpacing),
+          padding: EdgeInsetsDirectional.only(end: contentStyle.middleSpacing),
           child: Column(
             mainAxisSize: MainAxisSize.min,
             mainAxisAlignment: MainAxisAlignment.center,
             crossAxisAlignment: CrossAxisAlignment.start,
+            spacing: contentStyle.titleSpacing,
             children: [
               DefaultTextStyle.merge(
                 style: contentStyle.titleTextStyle.resolve(states),
@@ -80,17 +72,14 @@ class FTileContent extends StatelessWidget {
                 child: title,
               ),
               if (subtitle case final subtitle?)
-                Padding(
-                  padding: EdgeInsets.only(top: contentStyle.titleSpacing),
-                  child: DefaultTextStyle.merge(
-                    style: contentStyle.subtitleTextStyle.resolve(states),
-                    textHeightBehavior: const TextHeightBehavior(
-                      applyHeightToFirstAscent: false,
-                      applyHeightToLastDescent: false,
-                    ),
-                    overflow: TextOverflow.ellipsis,
-                    child: subtitle,
+                DefaultTextStyle.merge(
+                  style: contentStyle.subtitleTextStyle.resolve(states),
+                  textHeightBehavior: const TextHeightBehavior(
+                    applyHeightToFirstAscent: false,
+                    applyHeightToLastDescent: false,
                   ),
+                  overflow: TextOverflow.ellipsis,
+                  child: subtitle,
                 ),
             ],
           ),
@@ -109,16 +98,154 @@ class FTileContent extends StatelessWidget {
           const SizedBox(),
         if (suffixIcon case final suffixIcon?)
           Padding(
-            padding: ltr
-                ? EdgeInsets.only(left: contentStyle.suffixIconSpacing)
-                : EdgeInsets.only(right: contentStyle.suffixIconSpacing),
+            padding: EdgeInsetsDirectional.only(start: contentStyle.suffixIconSpacing),
             child: IconTheme(data: contentStyle.suffixIconStyle.resolve(states), child: suffixIcon),
           )
         else
           const SizedBox(),
-        if (dividerType != FTileDivider.none) FDivider(style: dividerStyle.resolve(states)) else const SizedBox(),
+        if (dividerType == FTileDivider.none) const SizedBox() else FDivider(style: dividerStyle.resolve(states)),
       ],
     );
+  }
+}
+
+class _TileContent extends MultiChildRenderObjectWidget {
+  final FTileContentStyle style;
+  final FTileDivider divider;
+
+  const _TileContent({required this.style, required this.divider, super.children});
+
+  @override
+  RenderObject createRenderObject(BuildContext context) =>
+      _RenderTileContent(style, divider, Directionality.maybeOf(context) ?? TextDirection.ltr);
+
+  @override
+  void updateRenderObject(BuildContext context, covariant _RenderTileContent content) {
+    content
+      ..style = style
+      ..divider = divider
+      ..textDirection = Directionality.maybeOf(context) ?? TextDirection.ltr;
+  }
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties
+      ..add(DiagnosticsProperty('style', style))
+      ..add(EnumProperty('divider', divider));
+  }
+}
+
+class _RenderTileContent extends RenderBox
+    with ContainerRenderObjectMixin<RenderBox, DefaultData>, RenderBoxContainerDefaultsMixin<RenderBox, DefaultData> {
+  FTileContentStyle _style;
+  FTileDivider _divider;
+  TextDirection _textDirection;
+
+  _RenderTileContent(this._style, this._divider, this._textDirection);
+
+  @override
+  void setupParentData(covariant RenderObject child) => child.parentData = DefaultData();
+
+  @override
+  void performLayout() {
+    final EdgeInsets(:left, :top, :right, :bottom) = _style.padding.resolve(_textDirection);
+    final prefix = firstChild!;
+    final column = childAfter(prefix)!;
+    final details = childAfter(column)!;
+    final suffix = childAfter(details)!;
+    final divider = childAfter(suffix)!;
+
+    // Layout children.
+    var contentConstraints = constraints.loosen().copyWith(maxWidth: constraints.maxWidth - left - right);
+
+    prefix.layout(contentConstraints, parentUsesSize: true);
+    contentConstraints = contentConstraints.copyWith(maxWidth: contentConstraints.maxWidth - prefix.size.width);
+
+    suffix.layout(contentConstraints, parentUsesSize: true);
+    contentConstraints = contentConstraints.copyWith(maxWidth: contentConstraints.maxWidth - suffix.size.width);
+
+    // Column takes priority if details is text, and vice-versa.
+    final (first, last) = details is RenderParagraph ? (column, details) : (details, column);
+
+    first.layout(contentConstraints, parentUsesSize: true);
+    contentConstraints = contentConstraints.copyWith(maxWidth: contentConstraints.maxWidth - first.size.width);
+    last.layout(contentConstraints, parentUsesSize: true);
+
+    // Layout divider based on the type.
+    switch (_divider) {
+      case FTileDivider.none || FTileDivider.full:
+        divider.layout(constraints.loosen(), parentUsesSize: true);
+
+      case FTileDivider.indented:
+        final spacing = _textDirection == TextDirection.ltr ? left : right;
+        final width = constraints.maxWidth - spacing - prefix.size.width;
+        divider.layout(constraints.loosen().copyWith(maxWidth: width), parentUsesSize: true);
+    }
+
+    final height = [prefix.size.height, suffix.size.height, column.size.height, details.size.height].max;
+    size = Size(constraints.maxWidth, height + top + bottom);
+
+    // Position children.
+    final (l, ml, mr, r) = _textDirection == TextDirection.ltr
+        ? (prefix, column, details, suffix)
+        : (suffix, details, column, prefix);
+
+    l.data.offset = Offset(left, top + (height - l.size.height) / 2);
+    ml.data.offset = Offset(left + l.size.width, top + (height - ml.size.height) / 2);
+    mr.data.offset = Offset(
+      constraints.maxWidth - right - r.size.width - mr.size.width,
+      top + (height - mr.size.height) / 2,
+    );
+    r.data.offset = Offset(constraints.maxWidth - right - r.size.width, top + (height - r.size.height) / 2);
+
+    divider.data.offset = Offset(
+      textDirection == TextDirection.ltr && _divider == FTileDivider.indented ? left + prefix.size.width : 0,
+      top + height + bottom - divider.size.height,
+    );
+  }
+
+  @override
+  void paint(PaintingContext context, Offset offset) => defaultPaint(context, offset);
+
+  @override
+  bool hitTestChildren(BoxHitTestResult result, {required Offset position}) =>
+      defaultHitTestChildren(result, position: position);
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties
+      ..add(DiagnosticsProperty('style', style))
+      ..add(EnumProperty('divider', divider))
+      ..add(EnumProperty('textDirection', textDirection));
+  }
+
+  FTileContentStyle get style => _style;
+
+  set style(FTileContentStyle value) {
+    if (_style != value) {
+      _style = value;
+      markNeedsLayout();
+    }
+  }
+
+  FTileDivider get divider => _divider;
+
+  set divider(FTileDivider value) {
+    if (_divider != value) {
+      _divider = value;
+      markNeedsLayout();
+    }
+  }
+
+  TextDirection get textDirection => _textDirection;
+
+  set textDirection(TextDirection value) {
+    if (_textDirection != value) {
+      _textDirection = value;
+      markNeedsLayout();
+    }
   }
 }
